Login: Extract footer links into their own component
